refactor: segment encoder class fakes for more robust test setup
